requirements.txt
bcpandas==0.7.1
mysql-connector==2.2.9
numpy==1.19.0
pandas==1.0.5
pyodbc==4.0.30
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
utils==1.0.1
